requirements.txt
discord.py
plexapi